got all the tests/err reporting to work
