+++ .github/workflows/check.yml
@@ -41,3 +41,7 @@ jobs:
       - name: Check Enclave Compile
         run: |
           yarn enclave:ci
+      
+      - name: Check Mock-Enclave VM Compile
+        run: |
+          yarn mock-enclave:test
