--- mock-enclave/result.json
@@ -1,5 +1,0 @@
-{
-"outcome": "None",
-"state": "{\"YgkAAABzeXN0ZW0uc2s=\":\"DAAAAHN5c3RlbV9oZWxsbw==\",\"U1RBVEU=\":\"AQAAAGI=\"}",
-"error": "None"
-}
