+++ mock-enclave/src/near-test-contracts/src/lib.rs
@@ -21,25 +21,25 @@ pub fn rs_contract() -> &'static [u8] {
     CONTRACT.get_or_init(|| read_contract("test_contract_rs.wasm")).as_slice()
 }
 
-pub fn rs_contract_base_protocol() -> &'static [u8] {
-    static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
-    CONTRACT.get_or_init(|| read_contract("test_contract_rs_base_protocol.wasm")).as_slice()
-}
+// pub fn rs_contract_base_protocol() -> &'static [u8] {
+//     static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
+//     CONTRACT.get_or_init(|| read_contract("test_contract_rs_base_protocol.wasm")).as_slice()
+// }
 
-pub fn nightly_rs_contract() -> &'static [u8] {
-    static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
-    CONTRACT.get_or_init(|| read_contract("nightly_test_contract_rs.wasm")).as_slice()
-}
+// pub fn nightly_rs_contract() -> &'static [u8] {
+//     static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
+//     CONTRACT.get_or_init(|| read_contract("nightly_test_contract_rs.wasm")).as_slice()
+// }
 
 pub fn ts_contract() -> &'static [u8] {
     static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
     CONTRACT.get_or_init(|| read_contract("test_contract_ts.wasm")).as_slice()
 }
 
-pub fn fuzzing_contract() -> &'static [u8] {
-    static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
-    CONTRACT.get_or_init(|| read_contract("contract_for_fuzzing_rs.wasm")).as_slice()
-}
+// pub fn fuzzing_contract() -> &'static [u8] {
+//     static CONTRACT: OnceCell<Vec<u8>> = OnceCell::new();
+//     CONTRACT.get_or_init(|| read_contract("contract_for_fuzzing_rs.wasm")).as_slice()
+// }
 
 /// Read given wasm file or panic if unable to.
 fn read_contract(file_name: &str) -> Vec<u8> {
@@ -54,11 +54,11 @@ fn read_contract(file_name: &str) -> Vec<u8> {
 #[test]
 fn smoke_test() {
     assert!(!rs_contract().is_empty());
-    assert!(!nightly_rs_contract().is_empty());
+    // assert!(!nightly_rs_contract().is_empty());
     assert!(!ts_contract().is_empty());
     assert!(!trivial_contract().is_empty());
-    assert!(!fuzzing_contract().is_empty());
-    assert!(!rs_contract_base_protocol().is_empty());
+    // assert!(!fuzzing_contract().is_empty());
+    // assert!(!rs_contract_base_protocol().is_empty());
 }
 
 pub fn many_functions_contract(function_count: u32) -> Vec<u8> {
