+++ mock-enclave/src/skw-vm-engine/Cargo.toml
@@ -14,32 +14,25 @@ This crate implements the specification of the interface that SkyeKiwi Network s
 """
 
 [dependencies]
-borsh = "0.9"
 serde = { version = "1", features = ["derive"] }
 wasmparser = "0.78"
-memoffset = "0.6"
-pwasm-utils = "0.12"
-parity-wasm = "0.41"
-wasmi = "0.10.0"
+pwasm-utils = "0.18.2"
+parity-wasm = "0.42.2"
 lru = "0.7.2"
 
-anyhow = { version = "1.0.19", optional = true }
+wasmi = { git = "https://github.com/paritytech/wasmi", features = ['v1'] }
 
 skw-vm-host = { path = "../skw-vm-host", default-features = false, features = [] }
 skw-vm-primitives = { path = "../skw-vm-primitives" }
-
 lazy_static = "1.4.0"
-near-primitives = { git = "https://github.com/near/nearcore", tag = "crates-0.11.0" }
 
-cached = "0.23.0"
 tracing = { version = "0.1", default-features = false }
-threadpool = "1.8.1"
 
 [dev-dependencies]
 near-test-contracts = { path = "../near-test-contracts" }
 assert_matches = "1.3"
 wat = "1.0.40"
-base64 = "0.13"
+
 
 [features]
 default = []
