+++ mock-enclave/src/skw-vm-engine/src/lib.rs
@@ -1,8 +1,6 @@
 mod cache;
 pub mod prepare;
 mod runner;
-mod externals;
-mod imports;
 
 #[cfg(test)]
 mod tests;
