+++ mock-enclave/src/skw-vm-engine/src/runner.rs
@@ -2,20 +2,23 @@ use crate::{cache};
 
 use skw_vm_primitives::contract_runtime::{ContractCode};
 use skw_vm_primitives::fees::RuntimeFeesConfig;
-use crate::externals::VMHost;
 
 use skw_vm_primitives::errors::{
-    FunctionCallError, MethodResolveError, VMError, WasmTrap,
+    FunctionCallError, MethodResolveError, VMError, WasmTrap, HostError,
 };
 use skw_vm_host::types::{PromiseResult};
 use skw_vm_host::{MemoryLike, VMConfig, VMContext, VMLogic, VMOutcome, RuntimeExternal};
 
 use wasmi::{
     MemoryInstance, MemoryRef, ModuleInstance,
-    memory_units::{Pages, Bytes, size_of}
+    memory_units::{Pages, Bytes, size_of},
+    Trap, TrapCode,
 };
 
-pub struct WasmiMemory(MemoryRef);
+// impl WasmiHostError for HostError {}
+
+#[derive(Clone)]
+pub struct WasmiMemory(pub MemoryRef);
 
 impl WasmiMemory {
     pub fn new(
@@ -28,6 +31,10 @@ impl WasmiMemory {
             )
             .expect("TODO creating memory cannot fail"),
         ))
+    }
+
+    pub fn clone(&self) -> MemoryRef {
+        self.0.clone()
     }
 }
 
@@ -82,6 +89,36 @@ fn check_method(module: &ModuleInstance, method_name: &str) -> Result<(), VMErro
     }
 }
 
+fn map_invoke_err(err: wasmi::Error) -> VMError {
+
+    println!("Real {:?}", err);
+
+    // Real Trap(Host(HostError(GasExceeded)))
+    let result = match err {
+        wasmi::Error::Trap(Trap::Code(e)) => {
+            match e {
+                TrapCode::Unreachable => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::Unreachable)),
+                TrapCode::MemoryAccessOutOfBounds => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::MemoryOutOfBounds)),
+                TrapCode::TableAccessOutOfBounds => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::MemoryOutOfBounds)),
+                TrapCode::ElemUninitialized => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::IndirectCallToNull)),
+                TrapCode::DivisionByZero => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::IllegalArithmetic)),
+                TrapCode::IntegerOverflow => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::IllegalArithmetic)),
+                TrapCode::InvalidConversionToInt => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::IllegalArithmetic)),
+                TrapCode::StackOverflow => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::StackOverflow)),
+                TrapCode::UnexpectedSignature => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::IncorrectCallIndirectSignature)),
+            }
+        },
+        wasmi::Error::Trap(Trap::Host(e)) => {
+            println!("{:?}", e.downcast_ref::<HostError>());
+            println!("before {:?}", e);
+            VMError::FunctionCallError(FunctionCallError::HostError(e.downcast_ref::<HostError>().unwrap().clone()))
+        },
+        _ => VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::Unreachable))
+    };
+    println!("Fake {:?}", result.clone());
+    result
+}
+
 pub struct WasmiVM;
 impl WasmiVM {
     pub fn run(
@@ -101,7 +138,42 @@ impl WasmiVM {
         )
         .entered();
 
-        let module = cache::create_module_instance(&code, wasm_config);
+        if method_name.is_empty() {
+            return (
+                None,
+                Some(VMError::FunctionCallError(FunctionCallError::MethodResolveError(
+                    MethodResolveError::MethodEmptyName,
+                ))),
+            );
+        }
+
+        let mut memory = WasmiMemory::new(
+            wasm_config.limit_config.initial_memory_pages,
+            wasm_config.limit_config.max_memory_pages,
+        ).expect("Cannot create memory for a contract call");
+        let memory_copy = memory.clone();
+
+        let mut logic = VMLogic::new(
+            ext, 
+            context, 
+            wasm_config, 
+            fees_config, 
+            promise_results, 
+            &mut memory
+        );
+
+        if logic.add_contract_compile_fee(code.code.len() as u64).is_err() {
+            return (
+                Some(logic.outcome()),
+                Some(VMError::FunctionCallError(FunctionCallError::HostError(
+                    HostError::GasExceeded
+                )))
+            )
+        }
+
+        println!("{:?}", logic.clone_outcome());
+
+        let module = cache::create_module_instance(&code, wasm_config, memory_copy);
 
         let module = match module {
             Ok(m) => m,
@@ -109,28 +181,18 @@ impl WasmiVM {
                 return (None, Some(VMError::FunctionCallError(FunctionCallError::CompilationError(e))));
             }
         };
-
+        
         if let Err(e) = check_method(&module, method_name) {
             return (None, Some(e));
         }
 
-
-        let mut memory = WasmiMemory::new(
-            wasm_config.limit_config.initial_memory_pages,
-            wasm_config.limit_config.max_memory_pages,
-        )
-        .expect("Cannot create memory for a contract call");
-
-        let logic = VMLogic::new(ext, context, wasm_config, fees_config, promise_results, &mut memory);
-        let mut host = VMHost(logic);
-
         let result = module
-            .invoke_export(&method_name, &[], &mut host)
-            .map_err(|_| FunctionCallError::WasmTrap(WasmTrap::MemoryOutOfBounds));
+            .invoke_export(&method_name, &[], &mut logic)
+            .map_err(map_invoke_err);
         
-        (Some(host.0.outcome()), match result {
+        (Some(logic.outcome()), match result {
             Ok(_) => None,
-            Err(_) => Some(VMError::FunctionCallError(FunctionCallError::WasmTrap(WasmTrap::MemoryOutOfBounds)))
+            Err(e) => Some(e)
         })
     }
 }
