+++ mock-enclave/src/skw-vm-host/Cargo.toml
@@ -14,22 +14,18 @@ This crate implements the specification of the interface that Near blockchain ex
 
 [dependencies]
 base64 = "0.13"
-borsh = "0.9"
 bs58 = "0.4"
 byteorder = "1.2"
 ripemd160 = "0.9.0"
 serde = { version = "1", features = ["derive"] }
 sha2 = ">=0.8,<0.10"
 sha3 = ">=0.8,<0.10"
-parity-secp256k1 = "0.7"
 
+wasmi = { git = "https://github.com/paritytech/wasmi", features = ['v1'] }
 near-crypto = { git = "https://github.com/near/nearcore" }
 skw-vm-primitives = { path = "../skw-vm-primitives" }
 
-bn = { package = "zeropool-bn", version = "0.5.9", features = [], optional = true }
-
 [dev-dependencies]
-hex = { version = "0.4", features = ["serde"] }
 serde_json = { version = "1", features = ["preserve_order"] }
 
 [features]
