+++ mock-enclave/src/skw-vm-host/src/lib.rs
@@ -1,6 +1,8 @@
 mod context;
 mod dependencies;
 mod logic;
+mod externals;
+mod imports;
 pub mod mocks;
 pub mod serde_with;
 pub mod types;
@@ -8,11 +10,11 @@ pub mod gas_counter;
 pub use context::VMContext;
 pub use dependencies::{RuntimeExternal, MemoryLike, ValuePtr};
 pub use logic::{VMLogic, VMOutcome};
+pub use imports::{WasmiImportResolver, create_builder};
 
 pub use skw_vm_primitives::config::*;
 pub use skw_vm_primitives::profile;
 pub use skw_vm_primitives::contract_runtime::ProtocolVersion;
-
 pub use skw_vm_primitives::errors::{HostError, VMLogicError};
 pub use types::ReturnData;
